requirements.txt
fastapi
uvicorn
newspaper3k
openai
langchain
pyttsx3
SpeechRecognition
twilio
pydantic
langchain_community
langchain-community
python-dotenv
langchain-core
langgraph>0.2.27
langchain-openai
lxml[html_clean]
langchain-text-splitters
pypdf
faiss-cpu
IPython
pydantic-settings
fastapi[all]
langsmith